requirements-tests.txt
-r requirements.txt
pytest-mock
pytest-celery
